@@ -16,13 +16,13 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~13.2.2",
-    "@angular/core": "~13.2.2",
+    "@angular/common": "^14.3.0",
+    "@angular/core": "^14.3.0",
     "@angular/fire": "^7.3.0",
-    "@angular/forms": "~13.2.2",
-    "@angular/platform-browser": "~13.2.2",
-    "@angular/platform-browser-dynamic": "~13.2.2",
-    "@angular/router": "~13.2.2",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@awesome-cordova-plugins/pdf-generator": "^6.2.0",
     "@capacitor-community/app-icon": "^3.1.0-beta.0",
     "@capacitor-community/http": "^1.4.1",
@@ -46,15 +46,15 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.3",
+    "@angular-devkit/build-angular": "^14.2.13",
     "@angular-eslint/builder": "~13.0.1",
     "@angular-eslint/eslint-plugin": "~13.0.1",
     "@angular-eslint/eslint-plugin-template": "~13.0.1",
     "@angular-eslint/template-parser": "~13.0.1",
-    "@angular/cli": "~13.2.3",
-    "@angular/compiler": "~13.2.2",
-    "@angular/compiler-cli": "~13.2.2",
-    "@angular/language-service": "~13.2.2",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler": "^14.3.0",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@capacitor/cli": "^4.0.0",
     "@ionic/angular-toolkit": "^6.0.0",
     "@types/jasmine": "~3.6.0",
@@ -76,7 +76,7 @@
     "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
-    "typescript": "~4.4.4"
+    "typescript": "~4.8.4"
   },
   "description": "An Ionic project"
 }
